WIP: Moving routes to separate files
