+++ routes/choreRoutes.js
@@ -1,7 +1,22 @@
 const Router = require('../modules/router');
+const loader = require('../modules/dataLoader');
+const helpers = require('../modules/routeHelpers');
+
+const Chore = require('../modules/sharedLibs/chore');
+
 let router = module.exports = new Router();
 
+router.add('GET', /^\/all-chores$/, function (request, response) {
+  helper.simpleGET(loader.allChores(), response);
+});
+
+router.add('GET', /^\/chore\/(\w+)$/, function (request, response, id) {
+  if (!id) respondBadReq(response, 'request must contain chore ID in path');
+  else simpleGET(loader.getChore(id), response);
+});
+
 router.add('PUT', /^\/chore$/, function (request, response) {
+  console.log("calling put chore!");
   const body = request.body;
   const invalidMsg = validateRequest(body, ['name', 'amount']);
   if (invalidMsg) {
